feat: 저장된 이미지 RecyclerView에 표시

+++ app/src/main/java/com/example/imageanalyzer/Picture.kt
@@ -2,4 +2,4 @@ package com.example.imageanalyzer
 
 import android.graphics.Bitmap
 
-data class Picture(val id:Int, val uri:String)
+data class Picture(val id:Long, val uri:String)
